Removing incorrect primary key requirement
